+++ src/de/omegazirkel/risingworld/tools/ui/DropdownCallback.java
@@ -1,0 +1,6 @@
+package de.omegazirkel.risingworld.tools.ui;
+
+@FunctionalInterface
+public interface DropdownCallback {
+    void onSelect(String key);
+}